Move duplicated DomainObjectIdentifier to the same package
