+++ subprojects/platform-base/src/main/java/dev/nokee/model/internal/DefaultNamedDomainObjectIdentifier.java
@@ -1,4 +1,4 @@
-package dev.nokee.platform.base.internal;
+package dev.nokee.model.internal;
 
 import lombok.Value;
