+++ subprojects/platform-base/src/main/java/dev/nokee/model/internal/NamedDomainObjectIdentifier.java
@@ -1,4 +1,6 @@
-package dev.nokee.platform.base.internal;
+package dev.nokee.model.internal;
+
+import dev.nokee.model.DomainObjectIdentifier;
 
 public interface NamedDomainObjectIdentifier extends DomainObjectIdentifier {
 	String getName();
