+++ subprojects/platform-base/src/main/java/dev/nokee/platform/base/DomainObjectCollection.java
@@ -1,6 +1,6 @@
 package dev.nokee.platform.base;
 
-import dev.nokee.platform.base.internal.DomainObjectIdentifier;
+import dev.nokee.model.DomainObjectIdentifier;
 import org.gradle.api.Transformer;
 import org.gradle.api.provider.Provider;
 import org.gradle.api.specs.Spec;
