+++ subprojects/platform-base/src/main/java/dev/nokee/platform/base/internal/DomainObjectFactory.java
@@ -1,5 +1,7 @@
 package dev.nokee.platform.base.internal;
 
+import dev.nokee.model.DomainObjectIdentifier;
+
 public interface DomainObjectFactory<T> {
 	T create();
