+++ subprojects/platform-base/src/main/java/dev/nokee/platform/base/internal/KnownDomainObjects.java
@@ -1,5 +1,6 @@
 package dev.nokee.platform.base.internal;
 
+import dev.nokee.model.DomainObjectIdentifier;
 import dev.nokee.platform.base.DomainObjectElement;
 import dev.nokee.platform.base.KnownDomainObject;
 import lombok.Getter;
