+++ subprojects/testing-base/src/main/java/dev/nokee/testing/base/internal/DefaultTestSuiteContainer.java
@@ -3,7 +3,7 @@ package dev.nokee.testing.base.internal;
 import dev.nokee.platform.base.DomainObjectProvider;
 import dev.nokee.platform.base.internal.AbstractDomainObjectContainer;
 import dev.nokee.platform.base.internal.DomainObjectFactory;
-import dev.nokee.platform.base.internal.DomainObjectIdentifier;
+import dev.nokee.model.DomainObjectIdentifier;
 import dev.nokee.platform.base.internal.DomainObjectStore;
 import dev.nokee.testing.base.TestSuiteComponent;
 import dev.nokee.testing.base.TestSuiteContainer;
