+++ subprojects/testing-xctest/src/main/java/dev/nokee/testing/xctest/internal/DefaultUiTestXCTestTestSuiteComponent.java
@@ -3,6 +3,7 @@ package dev.nokee.testing.xctest.internal;
 import com.google.common.collect.ImmutableList;
 import dev.nokee.core.exec.CommandLineTool;
 import dev.nokee.core.exec.internal.PathAwareCommandLineTool;
+import dev.nokee.model.DomainObjectIdentifier;
 import dev.nokee.platform.base.internal.*;
 import dev.nokee.platform.ios.internal.SignedIosApplicationBundleInternal;
 import dev.nokee.platform.ios.tasks.internal.CreateIosApplicationBundleTask;
