+++ subprojects/testing-xctest/src/main/java/dev/nokee/testing/xctest/internal/DefaultUnitTestXCTestTestSuiteComponent.java
@@ -3,6 +3,7 @@ package dev.nokee.testing.xctest.internal;
 import com.google.common.collect.ImmutableList;
 import dev.nokee.core.exec.CommandLineTool;
 import dev.nokee.core.exec.internal.PathAwareCommandLineTool;
+import dev.nokee.model.DomainObjectIdentifier;
 import dev.nokee.platform.base.internal.*;
 import dev.nokee.platform.ios.internal.IosApplicationBundleInternal;
 import dev.nokee.platform.ios.internal.SignedIosApplicationBundleInternal;
